+++ cmp-android/dependencies/prodReleaseRuntimeClasspath.tree.txt
@@ -2760,6 +2760,7 @@
 |    |    |    +--- org.jetbrains.androidx.navigation:navigation-compose:2.9.2 (*)
 |    |    |    +--- org.jetbrains.kotlinx:kotlinx-collections-immutable:0.5.1 (*)
 |    |    |    +--- project :core:model (*)
+|    |    |    +--- project :core:store (*)
 |    |    |    +--- org.jetbrains.compose.ui:ui:1.11.1 (*)
 |    |    |    +--- org.jetbrains.compose.foundation:foundation:1.11.1 (*)
 |    |    |    +--- org.jetbrains.compose.material3:material3:1.9.0 (*)
